chore(core): refactor nx release config implementation
