+++ packages/nx/src/command-line/release/version.ts
@@ -21,18 +21,16 @@ import {
   createProjectGraphAsync,
   readProjectsConfigurationFromProjectGraph,
 } from '../../project-graph/project-graph';
-import { findMatchingProjects } from '../../utils/find-matching-projects';
 import { combineOptionsForGenerator } from '../../utils/params';
 import { parseGeneratorString } from '../generate/generate';
 import { getGeneratorInformation } from '../generate/generator-utils';
 import { VersionOptions } from './command-object';
-import { createNxReleaseConfig } from './config/config';
 import {
   CATCH_ALL_RELEASE_GROUP,
-  ReleaseGroup,
-  createReleaseGroups,
-  handleCreateReleaseGroupsError,
-} from './config/create-release-groups';
+  createNxReleaseConfig,
+  handleNxReleaseConfigError,
+} from './config/config';
+import { filterReleaseGroups } from './config/filter-release-groups';
 import { printDiff } from './utils/print-changes';
 import { isRelativeVersionKeyword } from './utils/semver';
 
@@ -59,89 +57,33 @@ export async function versionHandler(args: VersionOptions): Promise<void> {
   }
 
   // Apply default configuration to any optional user configuration
-  const nxReleaseConfig = createNxReleaseConfig(nxJson.release);
-  const releaseGroupsData = await createReleaseGroups(
+  const { error: configError, nxReleaseConfig } = await createNxReleaseConfig(
     projectGraph,
-    nxReleaseConfig.groups
+    nxJson.release,
+    'nx-release-publish'
   );
-  if (releaseGroupsData.error) {
-    return await handleCreateReleaseGroupsError(releaseGroupsData.error);
+  if (configError) {
+    return await handleNxReleaseConfigError(configError);
+  }
+
+  const {
+    error: filterError,
+    releaseGroups,
+    releaseGroupToFilteredProjects,
+  } = filterReleaseGroups(
+    projectGraph,
+    nxReleaseConfig,
+    args.projects,
+    args.groups
+  );
+  if (filterError) {
+    output.error(filterError);
+    process.exit(1);
   }
 
   const tree = new FsTree(workspaceRoot, args.verbose);
 
-  let { releaseGroups } = releaseGroupsData;
-
-  /**
-   * User is filtering to a subset of projects. We need to make sure that what they have provided can be reconciled
-   * against their configuration in terms of release groups and the ungroupedProjectsHandling option.
-   */
   if (args.projects?.length) {
-    const matchingProjectsForFilter = findMatchingProjects(
-      args.projects,
-      projectGraph.nodes
-    );
-
-    if (!matchingProjectsForFilter.length) {
-      output.error({
-        title: `Your --projects filter "${args.projects}" did not match any projects in the workspace`,
-      });
-      process.exit(1);
-    }
-
-    const filteredProjectToReleaseGroup = new Map<string, ReleaseGroup>();
-    const releaseGroupToFilteredProjects = new Map<ReleaseGroup, Set<string>>();
-
-    // Figure out which release groups, if any, that the filtered projects belong to so that we can resolve other config
-    for (const releaseGroup of releaseGroups) {
-      const matchingProjectsForReleaseGroup = findMatchingProjects(
-        releaseGroup.projects,
-        projectGraph.nodes
-      );
-      for (const matchingProject of matchingProjectsForFilter) {
-        if (matchingProjectsForReleaseGroup.includes(matchingProject)) {
-          filteredProjectToReleaseGroup.set(matchingProject, releaseGroup);
-          if (!releaseGroupToFilteredProjects.has(releaseGroup)) {
-            releaseGroupToFilteredProjects.set(releaseGroup, new Set());
-          }
-          releaseGroupToFilteredProjects.get(releaseGroup).add(matchingProject);
-        }
-      }
-    }
-
-    /**
-     * If there are release groups specified, each filtered project must match at least one release
-     * group, otherwise the command + config combination is invalid.
-     */
-    if (Object.keys(nxReleaseConfig.groups).length) {
-      const unmatchedProjects = matchingProjectsForFilter.filter(
-        (p) => !filteredProjectToReleaseGroup.has(p)
-      );
-      if (unmatchedProjects.length) {
-        output.error({
-          title: `The following projects which match your projects filter "${args.projects}" did not match any configured release groups:`,
-          bodyLines: unmatchedProjects.map((p) => `- ${p}`),
-        });
-        process.exit(1);
-      }
-    }
-
-    output.note({
-      title: `Your filter "${args.projects}" matched the following projects:`,
-      bodyLines: matchingProjectsForFilter.map((p) => {
-        const releaseGroupForProject = filteredProjectToReleaseGroup.get(p);
-        if (releaseGroupForProject.name === CATCH_ALL_RELEASE_GROUP) {
-          return `- ${p}`;
-        }
-        return `- ${p} (release group "${releaseGroupForProject.name}")`;
-      }),
-    });
-
-    // Filter the releaseGroups collection appropriately
-    releaseGroups = releaseGroups.filter((rg) =>
-      releaseGroupToFilteredProjects.has(rg)
-    );
-
     /**
      * Run semver versioning for all remaining release groups and filtered projects within them
      */
@@ -181,21 +123,6 @@ export async function versionHandler(args: VersionOptions): Promise<void> {
     printChanges(tree, !!args.dryRun);
 
     return process.exit(0);
-  }
-
-  /**
-   * The user is filtering by release group
-   */
-  if (args.groups?.length) {
-    releaseGroups = releaseGroups.filter((g) => args.groups?.includes(g.name));
-  }
-
-  // Should be an impossible state, as we should have explicitly handled any errors/invalid config by now
-  if (!releaseGroups.length) {
-    output.error({
-      title: `No projects could be matched for versioning, please report this case and include your nx.json config`,
-    });
-    process.exit(1);
   }
 
   /**
